python/requirements.txt
notebook==6.0.3
pyvmomi==6.7.3
boto3==1.12.21
PyYAML==5.3